mark as internals some traces
